requirements.txt
requests
pybit
python-binance